--- tests_mocked/conftest.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    Dummy conftest.py for actinia.
-
-    If you don't know what this is for, just leave it empty.
-    Read more about conftest.py under:
-    https://pytest.org/latest/plugins.html
-"""
-
-# import pytest
