--- tests_mocked/mock/actinia_raster_management_mock.py
@@ -1,391 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-#######
-# actinia-python-client is a python client for actinia - an open source REST
-# API for scalable, distributed, high performance processing of geographical
-# data that uses GRASS GIS for computational tasks.
-#
-# Copyright (c) 2022 mundialis GmbH & Co. KG
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-#######
-
-__license__ = "GPLv3"
-__author__ = "Anika Weinmann"
-__copyright__ = "Copyright 2022, mundialis GmbH & Co. KG"
-__maintainer__ = "Anika Weinmann"
-
-from .actinia_mock import (
-    ACTINIA_BASEURL,
-    ACTINIA_VERSION
-)
-
-
-get_rasters_mock = {
-    "accept_datetime": "2022-06-13 13:04:53.422064",
-    "accept_timestamp": 1655125493.4220603,
-    "api_info": {
-        "endpoint": "rasterlayersresource",
-        "method": "GET",
-        "path": f"/api/{ACTINIA_VERSION}/locations/nc_spm_08/mapsets/"
-        "PERMANENT/raster_layers",
-        "request_url": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/locations/"
-        "nc_spm_08/mapsets/PERMANENT/raster_layers",
-    },
-    "datetime": "2022-06-13 13:04:53.740702",
-    "http_code": 200,
-    "message": "Processing successfully finished",
-    "process_chain_list": [
-        {
-            "1": {
-                "inputs": {"mapset": "PERMANENT", "type": "raster"},
-                "module": "g.list",
-            }
-        }
-    ],
-    "process_log": [
-        {
-            "executable": "g.list",
-            "id": "1",
-            "parameter": ["mapset=PERMANENT", "type=raster"],
-            "return_code": 0,
-            "run_time": 0.10059475898742676,
-            "stderr": [""],
-            "stdout": "aspect\nbasin_50K\nboundary_county_500m\ncfactorbare_"
-            "1m\ncfactorgrow_1m\nel_D782_6m\nel_D783_6m\nel_D792_6m\nel_D793_"
-            "6m\nelev_lid792_1m\nelev_ned_30m\nelev_srtm_30m\nelev_state_500"
-            "m\nelevation\nelevation_shade\nfacility\ngeology_30m\nlakes"
-            "\nlandclass96\nlandcover_1m\nlanduse96_28m\nlsat7_2002_10\n"
-            "lsat7_2002_20\nlsat7_2002_30\nlsat7_2002_40\nlsat7_2002_50\n"
-            "lsat7_2002_61\nlsat7_2002_62\nlsat7_2002_70\nlsat7_2002_80"
-            "\nncmask_500m\northo_2001_t792_1m\nroadsmajor\nslope\nsoilsID\n"
-            "soils_Kfactor\nstreams_derived\ntowns\nurban\nzipcodes\n"
-            "zipcodes_dbl\n",
-        }
-    ],
-    "process_results": [
-        "aspect",
-        "basin_50K",
-        "boundary_county_500m",
-        "cfactorbare_1m",
-        "cfactorgrow_1m",
-        "el_D782_6m",
-        "el_D783_6m",
-        "el_D792_6m",
-        "el_D793_6m",
-        "elev_lid792_1m",
-        "elev_ned_30m",
-        "elev_srtm_30m",
-        "elev_state_500m",
-        "elevation",
-        "elevation_shade",
-        "facility",
-        "geology_30m",
-        "lakes",
-        "landclass96",
-        "landcover_1m",
-        "landuse96_28m",
-        "lsat7_2002_10",
-        "lsat7_2002_20",
-        "lsat7_2002_30",
-        "lsat7_2002_40",
-        "lsat7_2002_50",
-        "lsat7_2002_61",
-        "lsat7_2002_62",
-        "lsat7_2002_70",
-        "lsat7_2002_80",
-        "ncmask_500m",
-        "ortho_2001_t792_1m",
-        "roadsmajor",
-        "slope",
-        "soilsID",
-        "soils_Kfactor",
-        "streams_derived",
-        "towns",
-        "urban",
-        "zipcodes",
-        "zipcodes_dbl",
-    ],
-    "progress": {"num_of_steps": 1, "step": 1},
-    "resource_id": "resource_id-0a7cb1ed-4fab-4353-9aac-87bb292c5a48",
-    "status": "finished",
-    "time_delta": 0.31871747970581055,
-    "timestamp": 1655125493.7406285,
-    "urls": {
-        "resources": [],
-        "status": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/resources/testuser"
-        "/resource_id-0a7cb1ed-4fab-4353-9aac-87bb292c5a48",
-    },
-    "user_id": "testuser",
-}
-
-raster_info_resp = {
-    "cells": "2025000",
-    "cols": "1500",
-    "comments": '"int(zipcodes)"',
-    "creator": '"helena"',
-    "database": "/actinia_core/workspace/temp_db/"
-    "gisdbase_d6bc2ddef1fd4471a84261ea5b108417",
-    "datatype": "CELL",
-    "date": '"Sun Jan 14 22:36:04 2007"',
-    "description": '"generated by r.mapcalc"',
-    "east": "645000",
-    "ewres": "10",
-    "location": "nc_spm_08",
-    "map": "zipcodes",
-    "mapset": "PERMANENT",
-    "maptype": "raster",
-    "max": "27610",
-    "min": "27511",
-    "ncats": "27610",
-    "north": "228500",
-    "nsres": "10",
-    "rows": "1350",
-    "semantic_label": '"none"',
-    "source1": '""',
-    "source2": '""',
-    "south": "215000",
-    "timestamp": '"none"',
-    "title": '"South West Wake: Zipcode areas derived from vector map"',
-    "units": '"none"',
-    "vdatum": '"none"',
-    "west": "630000",
-}
-
-
-get_raster_info_mock = {
-    "accept_datetime": "2022-06-13 13:28:23.934771",
-    "accept_timestamp": 1655126903.934769,
-    "api_info": {
-        "endpoint": "rasterlayerresource",
-        "method": "GET",
-        "path": f"/api/{ACTINIA_VERSION}/locations/nc_spm_08/mapsets/PERMANENT"
-        "/raster_layers/zipcodes",
-        "request_url": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/locations/"
-        "nc_spm_08/mapsets/PERMANENT/raster_layers/zipcodes",
-    },
-    "datetime": "2022-06-13 13:28:24.685413",
-    "http_code": 200,
-    "message": "Processing successfully finished",
-    "process_chain_list": [
-        {
-            "1": {
-                "flags": "gre",
-                "inputs": {"map": "zipcodes@PERMANENT"},
-                "module": "r.info",
-            }
-        }
-    ],
-    "process_log": [
-        {
-            "executable": "r.info",
-            "id": "1",
-            "mapset_size": 421,
-            "parameter": ["map=zipcodes@PERMANENT", "-gre"],
-            "return_code": 0,
-            "run_time": 0.1002960205078125,
-            "stderr": [""],
-            "stdout": "north=228500\nsouth=215000\neast=645000\nwest=630000"
-            "\nnsres=10\newres=10\nrows=1350\ncols=1500\ncells=2025000\n"
-            "datatype=CELL\nncats=27610\nmin=27511\nmax=27610\nmap=zipcodes\n"
-            "maptype=raster\nmapset=PERMANENT\nlocation=nc_spm_08\ndatabase"
-            "=/actinia_core/workspace/temp_db/gisdbase_d6bc2ddef1fd4471a8426"
-            '1ea5b108417\ndate="Sun Jan 14 22:36:04 2007"\ncreator="helena"\n'
-            'title="South West Wake: Zipcode areas derived from vector map"\n'
-            'timestamp="none"\nunits="none"\nvdatum="none"\nsemantic_label='
-            '"none"\nsource1=""\nsource2=""\ndescription="generated by '
-            'r.mapcalc"\ncomments="int(zipcodes)"\n',
-        }
-    ],
-    "process_results": raster_info_resp,
-    "progress": {"num_of_steps": 1, "step": 1},
-    "resource_id": "resource_id-165e9eca-9ab1-488b-8f49-d8b61f02315c",
-    "status": "finished",
-    "time_delta": 0.7506742477416992,
-    "timestamp": 1655126904.6853848,
-    "urls": {
-        "resources": [],
-        "status": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/resources/testuser"
-        "/resource_id-165e9eca-9ab1-488b-8f49-d8b61f02315c",
-    },
-    "user_id": "testuser",
-}
-
-upload_raster_resp = {
-    "accept_datetime": "2022-06-13 17:15:38.438675",
-    "accept_timestamp": 1655140538.4386735,
-    "api_info": {
-        "endpoint": "rasterlayerresource",
-        "method": "POST",
-        "path": f"/api/{ACTINIA_VERSION}/locations/nc_spm_08/mapsets/"
-                "raster_upload/raster_layers/test_elevation",
-        "request_url": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/locations/"
-                       "nc_spm_08/mapsets/raster_upload/raster_layers/"
-                       "test_elevation",
-    },
-    "datetime": "2022-06-13 17:15:41.866644",
-    "http_code": 200,
-    "message": "Raster layer <test_elevation> successfully imported.",
-    "process_chain_list": [
-        {
-            "1": {
-                "inputs": {
-                    "mapset": "raster_upload",
-                    "pattern": "test_elevation",
-                    "type": "raster",
-                },
-                "module": "g.list",
-            }
-        },
-        {
-            "1": {
-                "inputs": {
-                    "input": "/actinia_core/workspace/download_cache/testuser/"
-                             "home_testuser_data_tif_elevation_8e03bc94-58fc-"
-                             "449d-a4d2-4f4a2ac326a6.tif",
-                    "output": "test_elevation",
-                },
-                "module": "r.import",
-            }
-        },
-    ],
-    "process_log": [
-        {
-            "executable": "g.list",
-            "id": "1",
-            "mapset_size": 489,
-            "parameter": [
-                "type=raster",
-                "pattern=test_elevation",
-                "mapset=raster_upload",
-            ],
-            "return_code": 0,
-            "run_time": 0.10027885437011719,
-            "stderr": [""],
-            "stdout": "",
-        },
-        {
-            "executable": "r.import",
-            "id": "1",
-            "mapset_size": 36527,
-            "parameter": [
-                "input=/actinia_core/workspace/download_cache/testuser/"
-                "home_testuser_data_tif_elevation_8e03bc94-58fc-449d-a4d2-"
-                "4f4a2ac326a6.tif",
-                "output=test_elevation",
-            ],
-            "return_code": 0,
-            "run_time": 0.9039969444274902,
-            "stderr": [
-                "Importing raster map <test_elevation>...",
-                "0..3..6..9..12..15..18..21..24..27..30..33..36..39..42..45.."
-                "48..51..54..57..60..63..66..69..72..75..78..81..84..87..90.."
-                "93..96..99..100",
-                "",
-            ],
-            "stdout": "",
-        },
-    ],
-    "process_results": {},
-    "progress": {"num_of_steps": 2, "step": 2},
-    "resource_id": "resource_id-512baa14-253b-4030-95f7-e6bd2aec7493",
-    "status": "finished",
-    "time_delta": 3.427985668182373,
-    "timestamp": 1655140541.8666322,
-    "urls": {
-        "resources": [],
-        "status": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/resources/testuser"
-                  "/resource_id-512baa14-253b-4030-95f7-e6bd2aec7493",
-    },
-    "user_id": "testuser",
-}
-
-start_job_resp = {
-    "accept_datetime": "2022-06-13 17:15:38.438675",
-    "accept_timestamp": 1655140538.4386735,
-    "api_info": {
-        "endpoint": "rasterlayerresource",
-        "method": "POST",
-        "path": f"/api/{ACTINIA_VERSION}/locations/nc_spm_08/mapsets/raster_"
-                "upload/raster_layers/test_elevation",
-        "request_url": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/locations/"
-                       "nc_spm_08/mapsets/raster_upload/raster_layers/"
-                       "test_elevation",
-    },
-    "datetime": "2022-06-13 17:15:38.446568",
-    "http_code": 200,
-    "message": "Resource accepted",
-    "process_chain_list": [],
-    "process_results": {},
-    "resource_id": "resource_id-512baa14-253b-4030-95f7-e6bd2aec7493",
-    "status": "accepted",
-    "time_delta": 0.007901191711425781,
-    "timestamp": 1655140538.4465668,
-    "urls": {
-        "resources": [],
-        "status": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/resources/testuser"
-                  "/resource_id-512baa14-253b-4030-95f7-e6bd2aec7493",
-    },
-    "user_id": "testuser",
-}
-
-delete_raster_resp = {
-    "accept_datetime": "2022-06-13 16:47:11.782701",
-    "accept_timestamp": 1655138831.7826996,
-    "api_info": {
-        "endpoint": "rasterlayerresource",
-        "method": "DELETE",
-        "path": f"/api/{ACTINIA_VERSION}/locations/nc_spm_08/mapsets/"
-        "raster_upload/raster_layers/test_elevation",
-        "request_url": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/locations/"
-        "nc_spm_08/mapsets/raster_upload/raster_layers/"
-        "test_elevation",
-    },
-    "datetime": "2022-06-13 16:47:12.590450",
-    "http_code": 200,
-    "message": "Raster layer <test_elevation> successfully removed.",
-    "process_chain_list": [
-        {
-            "1": {
-                "flags": "f",
-                "inputs": {"name": "test_elevation", "type": "raster"},
-                "module": "g.remove",
-            }
-        }
-    ],
-    "process_log": [
-        {
-            "executable": "g.remove",
-            "id": "1",
-            "parameter": ["type=raster", "name=test_elevation", "-f"],
-            "return_code": 0,
-            "run_time": 0.35126757621765137,
-            "stderr": ["Removing raster <test_elevation>", ""],
-            "stdout": "",
-        }
-    ],
-    "process_results": {},
-    "progress": {"num_of_steps": 1, "step": 1},
-    "resource_id": "resource_id-3f4ce1c1-8083-4658-ba0a-5af2e08c7b04",
-    "status": "finished",
-    "time_delta": 0.8077754974365234,
-    "timestamp": 1655138832.5904214,
-    "urls": {
-        "resources": [],
-        "status": f"{ACTINIA_BASEURL}/api/{ACTINIA_VERSION}/resources/testuser"
-        "/resource_id-3f4ce1c1-8083-4658-ba0a-5af2e08c7b04",
-    },
-    "user_id": "testuser",
-}
